docs: refactor "structure" example to be exemplary for v1
- rename contract file
